Remove CI badge for wings.pivotal.io
We had the same pipeline running on wings (concourse deployed via BOSH
on VMs) and hush-house (conourse deployed as a helm chart onto a GKE
cluster) for a while now. There were no noticable differences in regards
to nested containers. We now decided to remove the pipeline from wings
and just keep the one on hush-house.